Moves Passport Strategy Into Own File
